backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
pillow==11.3.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
djangorestframework-simplejwt==5.3.0
